rework vscode_condor_ssh logic for shell setup
Adapted to be more generic and detected the shell by actually testing that it works,
instead of relying on expected output e.g. from our containers.
This simplifies the logic and makes it more generic for easier application to other HTCondor systems.

This commit also changes the stdout: It becomes more silent, as unnecessary output is not forwarded to VSCode.